requirements.txt
gradio==5.24.0
ollama==0.6.1
openai-whisper==20250625
TTS==0.22.0
mcp==1.12.0
mysql-connector-python==9.0.0
torch
